refactor: move PolygonMap to own folder
