+++ src/PolygonDraw/PolygonDraw.tsx
@@ -1,11 +1,11 @@
-import React from 'react'
+import React from 'react';
 
-import { Coordinate } from 'types'
-import { createLeafletLatLngTupleFromCoordinate, ensurePolygonList } from '../helpers'
+import { Coordinate } from 'types';
+import { createLeafletLatLngTupleFromCoordinate, ensurePolygonList } from '../helpers';
 
-import { MAP } from '../constants'
-import { PolygonMap } from './PolygonMap/PolygonMap'
-import UndoRedoProvider, { usePolygonEditor } from './hooks/usePolygonEditor'
+import { MAP } from '../constants';
+import { PolygonMap } from '../PolygonMap/PolygonMap';
+import UndoRedoProvider, { usePolygonEditor } from './hooks/usePolygonEditor';
 
 export type Props<T extends Coordinate[] | Coordinate[][]> = {
     boundary?: Coordinate[];
