+++ src/PolygonMap/PolygonMap.tsx
@@ -3,15 +3,15 @@ import flatten from 'lodash.flatten';
 import { useCallback, useEffect, useRef, useState } from 'react';
 import { useMap } from 'react-leaflet';
 
-import { ActionBar } from '../../ActionBar/ActionBar';
-import { Modal } from '../../common/components/Modal';
-import { MAP } from '../../constants';
-import { ExportPolygonForm } from '../../conversion/ExportPolygonForm';
-import { ImportPolygonForm } from '../../conversion/ImportPolygonForm';
-import { createLeafletLatLngBoundsFromCoordinates } from '../../helpers';
-import { Container, Map } from '../../leaflet/Map';
-import { TileLayer } from '../../leaflet/TileLayer';
-import { Coordinate, RectangleSelection } from '../../types';
+import { ActionBar } from '../ActionBar/ActionBar';
+import { Modal } from '../common/components/Modal';
+import { MAP } from '../constants';
+import { ExportPolygonForm } from '../conversion/ExportPolygonForm';
+import { ImportPolygonForm } from '../conversion/ImportPolygonForm';
+import { createLeafletLatLngBoundsFromCoordinates } from '../helpers';
+import { Container, Map } from '../leaflet/Map';
+import { TileLayer } from '../leaflet/TileLayer';
+import { Coordinate, RectangleSelection } from '../types';
 import { BoundaryPolygon } from './components/BoundaryPolygon';
 import { ActivePolygon } from './components/ActivePolygon';
 import { InactivePolygon } from './components/InactivePolygon';
