+++ src/PolygonMap/components/Polygon.tsx
@@ -2,8 +2,8 @@ import { FunctionComponent } from 'react';
 import { Polygon as LeafletPolygon } from 'react-leaflet';
 
 import { Coordinate } from 'types';
-import { createLeafletLatLngFromCoordinate } from '../../../helpers';
-import { MAP } from '../../../constants';
+import { createLeafletLatLngFromCoordinate } from '../../helpers';
+import { MAP } from '../../constants';
 
 // refers to https://leafletjs.com/reference-1.6.0.html#path-dasharray and
 // https://developer.mozilla.org/de/docs/Web/SVG/Attribute/stroke-dasharray
