+++ src/PolygonMap/components/SelectionRectangle.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { Rectangle } from 'react-leaflet';
 import { LatLngBounds, latLngBounds } from 'leaflet';
 
-import { RectangleSelection } from '../../../types';
-import { MAP } from '../../../constants';
-import { createLeafletLatLngFromCoordinate } from '../../../helpers';
+import { RectangleSelection } from '../../types';
+import { MAP } from '../../constants';
+import { createLeafletLatLngFromCoordinate } from '../../helpers';
 
 interface Props {
     rectangleSelection: RectangleSelection;
